Improvements for playback
These improvements are needed to ensure that passing only the objects
that have been changed on the file system and using --no-recurse
--override will yield the same result as if everything is played back
(but faster).

* Execute the after_recurse_hook even if not recursing, allowing to play
back a change in the order of the items in an ordered folder
* If overriding an object due to a change in the meta type, force
recursion for the subtree. This allows to change between a "Folder" and
"Folder (Ordered)" and playing back that change
* Correct progress counting