Change handling of non-default cases in str_repr
Unicode in Python2 or Bytes in Python3 are now delegated to repr(),
leaving only the case of Bytes in Python2 to be handled differently. If
the bytes are not valid UTF-8, the method also falls back to repr().

Added more test cases

Rewrote docstring of str_repr to hopefully be more comprehensible.