Switch to using a single excutable `zodbsync` with subcommands
Current subpcommands are `record`, `playback`, `watch` and `pick`,
opening the way to add further functionality.

Remove unused features like recording since a given timestamp or
transaction ID. Also remove `--apply` functionality, which was not very
helpful in the case of merge conflicts. If you really need to apply a
patchfile, use `git am` directly, reset the branch to before the resulting
commit and `pick` it again.