Create testing environment
This creates an environment for developing further tests, including a
fixture that provides a full-fledged environment (running ZEO in a
temporary directory, repository and configuration files) that can be
used by tests. Two tests are included, which can be run by starting the
`tox` command in the checked out folder.