Refactor responsibilities of ZODBSync and Subcommand
ZODBSync now is only initialized if the command in question actually
needs a connection to the ZODB. Arguments in the initialization of
ZODBSync that were anyhow never used with non-default values were
dropped.

The argument parsing logic now happens in a new class Runner in main.py.
This cleans up an effective import circle (SubCommand.create was passed
a list of subclasses of SubCommand). It also prepares the way for
re-using the same ZODB connection for different commands, which should
simplify the tests somewhat.

Each command now releases the lock after execution.