Fix timing issues in watcher
Since I could not find a way to get the current transaction ID after
starting a transaction, lastTransaction() should be read before the
transaction is started. This ensures that we do not read an old state
and miss important changes in all further updates because we already
assumed we were reading the state at the transaction stored in
`self.last_visible_transaction`.

Fixes #73.