Call after-recurse-hook in correct order (#83)
If multiple paths are given to `playback_paths`, the after-recurse-hook
that tries to fix the ordering of children in the case of Ordered
Folders now is executed in the correct order. If A is the original
playback of the data of a given path and A' is the after-recurse-hook,
the operations are now executed A-B-C-C'-B'-A' instead of
A-A'-B-B'-C-C' on the sorted list of paths, ensuring the requested
ordering also for new children.